pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigilant-enigma1"
version = "0.1.0"
requires-python = ">=3.10"
description = "Minimal async web scraper with httpx/BeautifulSoup, optional Playwright, dotenv, tqdm, tenacity, CSV/pandas output."
authors = [{ name = "you" }]
dependencies = [
  "httpx==0.27.2",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "python-dotenv==1.0.1",
  "tqdm==4.66.5",
  "tenacity==9.0.0",
]

[project.optional-dependencies]
browser = ["playwright==1.47.0", "playwright-stealth>=1.0.6"]
data = ["pandas==2.2.2"]
all = ["playwright==1.47.0", "playwright-stealth>=1.0.6", "pandas==2.2.2"]

[project.scripts]
scrape = "vigilant_enigma1.cli:main"
igscrape = "igscraper.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["vigilant_enigma1*", "igscraper*"]
